Add abstract physics environment
